+++ app/client/scripts/pages/common.ts
@@ -1,5 +1,5 @@
 /**
- * @file ブラウザ関連処理の各ページの共通処理
+ * @file Common processes
  */
 import 'ts-polyfill/lib/es2015-core';
 import 'ts-polyfill/lib/es2015-promise';
@@ -9,10 +9,14 @@ import 'ts-polyfill/lib/es2017-string';
 import 'ts-polyfill/lib/es2017-object';
 import 'ts-polyfill/lib/es2018-promise';
 import WebFont from 'webfontloader';
+import picturefill from 'picturefill';
 
-// 指定されたWebフォントをロードする
+// polyfill
+picturefill();
+
+// Load web fonts
 WebFont.load({
   google: {
-    families: ['Droid Sans'], // TODO : ここにロードするWebフォントを記述する
+    families: ['Droid Sans'], // TODO : Specify using web fonts
   },
 });